move all the oshu_* symbols under oshu::
It bothered me to leave oshu after a half-assed C++ migrations. Some
things were inside the namespace and other had the oshu_ C prefix. Now
the naming scheme is at least consistent.

This refactoring was mostly automated, which is why oshu:: is explicit
in most instances. It doesn’t have to be that way in the future.